backport numeric keypad support from Joel Mawhorter
Joel Mawhorter <joel@mawhorter.org> wrote a patch to support
volume up/down for scroll up/down in a message view.  His
patch was against git master which has several problems
making it unusable on the sonim xp5.  So backport the patch
to 5.600, and make a few changes:

1. return true after handling 2/8.  As an attempt to see if
that makes it work
2. also use dpad up/down for scroll up/down.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>